refactor(kernel): untag pure-Go log pipeline and route drop warning off logger.Logger
Two review fixes:

- Move the pure-Go async forwarding pipeline (record type, bounded queue, drain,
  flush barrier, drop accounting, panic containment) out of the cgo-tagged
  log_callback.go into the untagged logforward_async.go, and its tests into
  logforward_async_test.go. This matches the repo convention documented in
  logging_level.go: the FIFO-flush, drop-policy, and panic-containment tests now run
  in the default CGO_ENABLED=0 lane instead of only the kernel-linked lane.
  log_callback.go keeps only the cgo trampoline and the one-time install.
- Route the drain's one-shot drop warning through the sink's own immutable logger
  (new logSink.warnDropped) instead of logger.Logger, whose embedded value
  SetLogLevel reassigns. The long-lived drain goroutine no longer reads a field that
  races SetLogLevel; drainKernelLogs is now pure Go.
